Fixed bugs, added rock wall, added props
Bug Fixes
Made it so enemies couldn’t spawn on top of the player any more.
Adjusted the UI to scale to the screen size.
Fixed all issues with the UIManager not working when the player dies or
goes to the next level.
Fixed problem with enemies traveling through walls

Rock Wall
Added rocks all the way around the map to attempt to hide the sharp
lines and make the cave look more natural.

Props
Added 4 new props, 2 sets of bones and 2 rocks. These sprites are
properly textured so they react to lighting.